Fix sidebar step prop and move sidebar components to new folder
