+++ src/lib/components/sidebar/SidebarMenuItemCircle.svelte
