+++ src/lib/components/sidebar/SidebarMenuItemText.svelte
